requirements.txt
websocket-client==1.6.1
pandas==2.0.3
apscheduler==3.10.1
python-dotenv==1.0.0
requests==2.31.0
pytz==2023.3
